tested qdnaseq performance when detecting ony one CNV: success

+++ workflow/rules/cnv_call.smk
@@ -18,7 +18,7 @@ rule run_qdnaseq:
         bins=config["cnv_binsize"]
     shell:
         """
-        Rscript workflow/scripts/qdnaseq_sop.R -t {threads} -b {params.bins} {input.aligned_unphased_bam} {params.output_trunk} > {log.o} 2> {log.e}
+        Rscript workflow/scripts/qdnaseq_sop_explicitPrint.R -t {threads} -b {params.bins} {input.aligned_unphased_bam} {params.output_trunk} > {log.o} 2> {log.e}
         """
 
 rule plot_qdnaseq:
@@ -57,21 +57,27 @@ rule annotate_cnvs:
         TRUNK={params.filetrunk}
         VCF={input.qdnaseq_vcf}
 
-        paste <( tail -n+2 $INPUTFILE | awk '{{print "chr"$2}}' ) <(tail -n+2 $INPUTFILE | cut -f3,4 ) > $TRUNK.intervals.bed
+        INLINES=$( wc -l $INPUTFILE | cut -d ' ' -f1)
+        if [[ $INLINES -lt 2 ]]
+        then
+            grep ^# $VCF > {output.qdnaseq_vcf}
+        else
+            paste <( tail -n+2 $INPUTFILE | awk '{{print "chr"$2}}' ) <(tail -n+2 $INPUTFILE | cut -f3,4 ) > $TRUNK.intervals.bed
 
-        bedtools intersect -b {params.genedf} -a $TRUNK.intervals.bed -wa -wb -loj > $TRUNK.gene.intersections.2.bed
+            bedtools intersect -b {params.genedf} -a $TRUNK.intervals.bed -wa -wb -loj > $TRUNK.gene.intersections.2.bed
 
-        grep ^## $VCF | head -n-1 > {output.qdnaseq_vcf}
-        echo "##INFO=<ID=GENECOUNT,Number=1,Type=Integer,Description=\"Number of overlapping genes in call\">" >> {output.qdnaseq_vcf}
-        echo "##INFO=<ID=GENENAMES,Number=1,Type=String,Description=\"Overlapping genes in call\">" >> {output.qdnaseq_vcf}
-        grep ^## $VCF | tail -n1 >> {output.qdnaseq_vcf}
-        grep ^#CHROM $VCF >> {output.qdnaseq_vcf}
+            grep ^## $VCF | head -n-1 > {output.qdnaseq_vcf}
+            echo "##INFO=<ID=GENECOUNT,Number=1,Type=Integer,Description=\"Number of overlapping genes in call\">" >> {output.qdnaseq_vcf}
+            echo "##INFO=<ID=GENENAMES,Number=1,Type=String,Description=\"Overlapping genes in call\">" >> {output.qdnaseq_vcf}
+            grep ^## $VCF | tail -n1 >> {output.qdnaseq_vcf}
+            grep ^#CHROM $VCF >> {output.qdnaseq_vcf}
 
-        paste <(paste <(grep -v ^# $VCF | cut -f1-7) <(paste <(grep -v ^# $VCF | cut -f8) <(paste <(awk 'NR==1{{id=$1":"$2;if($7=="."){{genes=0}}else{{genes=1}}}}\
-        NR>1{{newid=$1":"$2;if(id==newid){{genes+=1}}else{{print genes;id=newid;if($7=="."){{genes=0}}else{{genes=1}}}}}}END{{print genes}}' $TRUNK.gene.intersections.2.bed)\
-        <(awk 'NR==1{{id=$1":"$2;genes=$7}}NR>1{{newid=$1":"$2;if(id==newid){{genes=genes"-"$7}}else{{print genes;id=newid;genes=$7}}}}END{{print genes}}' $TRUNK.gene.intersections.2.bed) \
-        | awk '{{print "GENECOUNT="$1";GENENAMES="$2}}') | tr '\t' ';')) <(grep -v ^# $VCF | cut -f9,10) >> {output.qdnaseq_vcf}
+            paste <(paste <(grep -v ^# $VCF | cut -f1-7) <(paste <(grep -v ^# $VCF | cut -f8) <(paste <(awk 'NR==1{{id=$1":"$2;if($7=="."){{genes=0}}else{{genes=1}}}}\
+            NR>1{{newid=$1":"$2;if(id==newid){{genes+=1}}else{{print genes;id=newid;if($7=="."){{genes=0}}else{{genes=1}}}}}}END{{print genes}}' $TRUNK.gene.intersections.2.bed)\
+            <(awk 'NR==1{{id=$1":"$2;genes=$7}}NR>1{{newid=$1":"$2;if(id==newid){{genes=genes"-"$7}}else{{print genes;id=newid;genes=$7}}}}END{{print genes}}' $TRUNK.gene.intersections.2.bed) \
+            | awk '{{print "GENECOUNT="$1";GENENAMES="$2}}') | tr '\t' ';')) <(grep -v ^# $VCF | cut -f9,10) >> {output.qdnaseq_vcf}
 
-        rm -rf $TRUNK.gene.intersections.2.bed
-        rm -rf $TRUNK.intervals.bed
+            rm -rf $TRUNK.gene.intersections.2.bed
+            rm -rf $TRUNK.intervals.bed
+        fi
         """
